:recycle: refactor: `util` 패키지명 `exception`으로 변경
