+++ src/main/java/view/InputView.java
@@ -1,5 +1,6 @@
 package view;
 
+import exception.ErrorException;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
